Updated collection and book dao, added Message Timeout hook, updated Search component

+++ frontend/src/hooks/useFetch.jsx
@@ -1,11 +1,11 @@
-import { useState, useEffect, useContext } from "react";
+import React, { useState, useEffect, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import { AuthContext } from "../context/AuthProvider";
 
 
 export default function useFetch(url, options, refresh) {
-    const { setUsername, setRole, setToken, setTokenExpMessage } = useContext(AuthContext);
-    const navigate = useNavigate()
+    const { setUsername, setRole, setToken, setTokenExpMessage, token } = useContext(AuthContext);
+    const navigate = useNavigate();
     const [data, setData] = useState([]);
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState(null);
@@ -33,14 +33,13 @@ export default function useFetch(url, options, refresh) {
                     setData([]);
                 }
             } catch (error) {
-                console.log(error)
-                if (error.status == 401 || error.status == 403) {
-                    localStorage.clear()
-                    setUsername(null)
-                    setRole(null)
-                    setToken(null)
-                    setTokenExpMessage("Sign-in expired. Please sign-in again.")
-                    navigate(`/profile`)
+                if (token && (error.status == 401 || error.status == 403)) {
+                    localStorage.clear();
+                    setUsername(null);
+                    setRole(null);
+                    setToken(null);
+                    setTokenExpMessage("Sign-in expired. Please sign-in again.");
+                    navigate(`/profile`);
                 }
                 setError(error);
             } finally {
@@ -48,9 +47,8 @@ export default function useFetch(url, options, refresh) {
             }
         }
         fetchData();
-    }, [url, options, refresh, navigate, setUsername, setRole, setToken, setTokenExpMessage]);
+    }, [url, options, refresh, navigate, setUsername, setRole, setToken, setTokenExpMessage, token]);
 
-    console.log("fetch",data)
-    return { data, loading, error };
+    return { data, loading, error }
 }
